delete config nid convert error fix

+++ config/src/main/java/com/alibaba/nacos/config/server/service/repository/extrnal/ExternalHistoryConfigInfoPersistServiceImpl.java
@@ -44,6 +44,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.support.TransactionTemplate;
 
+import java.math.BigInteger;
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
@@ -92,13 +93,13 @@ public class ExternalHistoryConfigInfoPersistServiceImpl implements HistoryConfi
     public List<ConfigInfoWrapper> convertDeletedConfig(List<Map<String, Object>> list) {
         List<ConfigInfoWrapper> configs = new ArrayList<>();
         for (Map<String, Object> map : list) {
-            Long id = (Long) map.get("nid");
+            BigInteger id = (BigInteger) map.get("nid");
             String dataId = (String) map.get("data_id");
             String group = (String) map.get("group_id");
             String tenant = (String) map.get("tenant_id");
             long mTime = ((LocalDateTime) map.get("gmt_modified")).toInstant(ZoneOffset.ofHours(8)).toEpochMilli();
             ConfigInfoWrapper config = new ConfigInfoWrapper();
-            config.setId(id);
+            config.setId(id.longValue());
             config.setDataId(dataId);
             config.setGroup(group);
             config.setTenant(tenant);
